+++ common/src/main/java/dev/terminalmc/chatnotify/gui/widget/list/option/KeyOptionList.java
@@ -120,7 +120,7 @@ public class KeyOptionList extends OptionList {
                 triggerField.setResponder((str) -> trigger.string = str.strip());
                 triggerField.setValue(trigger.string);
                 triggerField.setTooltip(Tooltip.create(
-                        localized("option", "notif.trigger.field.tooltip")));
+                        localized("option", "trigger.field.tooltip")));
                 triggerField.setTooltipDelay(Duration.ofMillis(500));
                 elements.add(triggerField);
             }